combined pickle and dill balance cards and added total locked pickle amount to new card
